+++ app/Models/Partner.php
@@ -4,11 +4,19 @@ namespace App\Models;
 
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
+use Illuminate\Database\Eloquent\Relations\BelongsToMany;
+use Illuminate\Database\Eloquent\Relations\HasMany;
 use Illuminate\Database\Eloquent\SoftDeletes;
 
 class Partner extends Model
 {
     use HasFactory, SoftDeletes;
+
+    /*
+    |--------------------------------------------------------------------------
+    | GLOBAL VARIABLES
+    |--------------------------------------------------------------------------
+    */
 
     protected $fillable = ['name', 'phone', 'info', 'is_block', 'admin_user_id'];
 
@@ -19,9 +27,48 @@ class Partner extends Model
     const BLOCK_YES = 1; // Заблокирован
     const BLOCK_NO = 0; // Не заблокирован
 
-    public function partnerUsers()
+
+    /*
+    |--------------------------------------------------------------------------
+    | FUNCTIONS
+    |--------------------------------------------------------------------------
+    */
+
+    /*
+    |--------------------------------------------------------------------------
+    | RELATIONS
+    |--------------------------------------------------------------------------
+    */
+
+    public function partnerUsers1()
     {
         return $this->belongsToMany(User::class, 'partner_user')->withPivot('phone', 'setting_info', 'status');
+    }
+
+    /**
+     * @return \Illuminate\Database\Eloquent\Relations\BelongsToMany
+     */
+    public function employees() : BelongsToMany
+    {
+        /*return $this->belongsToMany(Partner::class)
+            ->using(PartnerUser::class)
+            ->withPivot('status', 'setting_info', 'is_admin');*/
+        return $this->belongsToMany(User::class)
+            ->using(PartnerUser::class)
+            ->withPivot('status', 'setting_info', 'is_admin');;
+    }
+
+    /**
+     * @return \Illuminate\Database\Eloquent\Relations\HasMany
+     */
+    public function partnerUsers() : hasMany
+    {
+        /*return $this->belongsToMany(Partner::class)
+            ->using(PartnerUser::class)
+            ->withPivot('status', 'setting_info', 'is_admin');*/
+        return $this->hasMany(PartnerUser::class);
+        //    ->withPivot('status', 'setting_info', 'is_admin');
+        //return $this->hasMany(PartnerUser::class);
     }
 
 
@@ -29,4 +76,37 @@ class Partner extends Model
     {
         return $this->belongsTo(User::class, 'admin_user_id', 'id');
     }
+
+    /*
+    |--------------------------------------------------------------------------
+    | SCOPES
+    |--------------------------------------------------------------------------
+    */
+
+    /**
+     * Заготовка запроса на получение списка партнеров через таблицу сотрудников для определенного пользователя
+     *
+     * @param \Illuminate\Database\Eloquent\Builder $query
+     * @param int $userId
+     * @return \Illuminate\Database\Eloquent\Builder
+     */
+    public function scopeConcretePartnerUser($query, $userId)
+    {
+        return $query->whereHas('partnerUsers', function ($query) use ($userId) {
+            $query->where('partner_user.user_id', $userId);
+        });
+    }
+
+
+    /*
+    |--------------------------------------------------------------------------
+    | ACCESSORS
+    |--------------------------------------------------------------------------
+    */
+
+    /*
+    |--------------------------------------------------------------------------
+    | MUTATORS
+    |--------------------------------------------------------------------------
+    */
 }