+++ app/Policies/AccountPolicy.php
@@ -1,0 +1,99 @@
+<?php
+
+namespace App\Policies;
+
+use App\JsonApi\Proxies\Account;
+use App\Models\User;
+use Illuminate\Auth\Access\HandlesAuthorization;
+
+class AccountPolicy
+{
+    use HandlesAuthorization;
+
+    /**
+     * Determine whether the user can view any models.
+     *
+     * @param  \App\Models\User  $user
+     * @return \Illuminate\Auth\Access\Response|bool
+     */
+    public function viewAny(User $user)
+    {
+        //return $user->isA('partner');
+        return true;
+    }
+
+    /**
+     * Determine whether the user can view the model.
+     *
+     * @param \App\Models\User $user
+     * @param Account $account
+     * @return \Illuminate\Auth\Access\Response|bool
+     */
+    public function view(User $user, Account $account)
+    {
+        //return $user->isA('partner');
+        return false;
+    }
+
+    /**
+     * Determine whether the user can create models.
+     *
+     * @param  \App\Models\User  $user
+     * @return \Illuminate\Auth\Access\Response|bool
+     */
+    public function create(User $user)
+    {
+        return true;
+    }
+
+    /**
+     * Determine whether the user can update the model.
+     *
+     * @param  \App\Models\User  $user
+     * @param  Account $account
+     * @return \Illuminate\Auth\Access\Response|bool
+     */
+    public function update(User $user, Account $account)
+    {
+        // и если есть право
+        // и если является админом
+        //return  $user->id == $partner->admin_user_id && $user->isA('partner', 'partnerAdmin');
+        return  false;
+    }
+
+    /**
+     * Determine whether the user can delete the model.
+     *
+     * @param  \App\Models\User  $user
+     * @return \Illuminate\Auth\Access\Response|bool
+     */
+    //public function delete(User $user, Account $account)
+    public function delete(User $user)
+    {
+        return true;
+    }
+
+    /**
+     * Determine whether the user can restore the model.
+     *
+     * @param  \App\Models\User  $user
+     * @param  Account $account
+     * @return \Illuminate\Auth\Access\Response|bool
+     */
+    public function restore(User $user, Account $account)
+    {
+        return  false;
+    }
+
+    /**
+     * Determine whether the user can permanently delete the model.
+     *
+     * @param  \App\Models\User  $user
+     * @param  Account $account
+     * @return \Illuminate\Auth\Access\Response|bool
+     */
+    public function forceDelete(User $user, Account $account)
+    {
+        return  false;
+    }
+}